Pluggable authentication. Google oauth as a module
