+++ engine/Dockerfile.interactive
@@ -10,8 +10,8 @@ MAINTAINER Tanmay Mohapatra
 EXPOSE 8888
 
 ADD www /jboxengine/www
-# Allow juser write permissions on admin_modules.tpl to generate its content at run time.
-RUN chmod 666 /jboxengine/www/admin_modules.tpl
+# Allow juser write permissions on admin_modules.tpl and auth_modules.tpl to generate its content at run time.
+RUN chmod 666 /jboxengine/www/admin_modules.tpl /jboxengine/www/auth_modules.tpl /jboxengine/www/session_modules.tpl
 
 USER juser
