+++ engine/src/juliabox/handlers/__init__.py
@@ -1,7 +1,6 @@
 __author__ = 'tan'
 from handler_base import JBoxHandler, JBoxHandlerPlugin, JBoxUIModulePlugin
 from admin import AdminHandler
-from auth import AuthHandler
 from cors import CorsHandler
 from main import MainHandler
 from ping import PingHandler
