+++ engine/src/juliabox/plugins/auth_google/__init__.py
@@ -1,0 +1,2 @@
+__author__ = 'tan'
+from google_auth import GoogleAuthHandler, GoogleAuthUIHandler
