+++ engine/src/juliabox/plugins/auth_google/google_login_btn.tpl
@@ -1,0 +1,1 @@
+<a class="btn btn-primary btn-block gauth-btn" title="Sign in via Google" href="/jboxauth/google/">Sign in via Google</a>
